[Security Solution][Alert details] - extract panel keys to constant file to avoid weird webpack issue (circular dependency?)
