+++ x-pack/plugins/security_solution/public/flyout/document_details/isolate_host/index.tsx
@@ -11,9 +11,6 @@ import type { FlyoutPanelProps } from '@kbn/expandable-flyout';
 import { PanelContent } from './content';
 import { PanelHeader } from './header';
 
-export const DocumentDetailsIsolateHostPanelKey: IsolateHostPanelProps['key'] =
-  'document-details-isolate-host';
-
 export interface IsolateHostPanelProps extends FlyoutPanelProps {
   key: 'document-details-isolate-host';
   params?: {
