+++ x-pack/plugins/security_solution/public/flyout/document_details/left/tabs/insights_tab.tsx
@@ -22,7 +22,8 @@ import {
   INSIGHTS_TAB_CORRELATIONS_BUTTON_TEST_ID,
 } from './test_ids';
 import { useLeftPanelContext } from '../context';
-import { DocumentDetailsLeftPanelKey, LeftPanelInsightsTab } from '..';
+import { DocumentDetailsLeftPanelKey } from '../../shared/constants/panel_keys';
+import { LeftPanelInsightsTab } from '..';
 import { ENTITIES_TAB_ID, EntitiesDetails } from '../components/entities_details';
 import {
   THREAT_INTELLIGENCE_TAB_ID,
