+++ x-pack/plugins/security_solution/public/flyout/document_details/left/tabs/visualize_tab.tsx
@@ -13,7 +13,8 @@ import { useExpandableFlyoutApi, useExpandableFlyoutState } from '@kbn/expandabl
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { useLeftPanelContext } from '../context';
-import { DocumentDetailsLeftPanelKey, LeftPanelVisualizeTab } from '..';
+import { DocumentDetailsLeftPanelKey } from '../../shared/constants/panel_keys';
+import { LeftPanelVisualizeTab } from '..';
 import {
   VISUALIZE_TAB_BUTTON_GROUP_TEST_ID,
   VISUALIZE_TAB_GRAPH_ANALYZER_BUTTON_TEST_ID,
