+++ x-pack/plugins/security_solution/public/flyout/document_details/right/components/alert_description.tsx
@@ -21,11 +21,8 @@ import {
   ALERT_DESCRIPTION_TITLE_TEST_ID,
   RULE_SUMMARY_BUTTON_TEST_ID,
 } from './test_ids';
-import {
-  DocumentDetailsPreviewPanelKey,
-  type PreviewPanelProps,
-  RulePreviewPanel,
-} from '../../preview';
+import { DocumentDetailsPreviewPanelKey } from '../../shared/constants/panel_keys';
+import { type PreviewPanelProps, RulePreviewPanel } from '../../preview';
 
 /**
  * Displays the rule description of a signal document.
