+++ x-pack/plugins/security_solution/public/flyout/document_details/right/navigation.tsx
@@ -11,7 +11,7 @@ import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
 import { useKibana } from '../../../common/lib/kibana';
 import { HeaderActions } from './components/header_actions';
 import { FlyoutNavigation } from '../../shared/components/flyout_navigation';
-import { DocumentDetailsLeftPanelKey } from '../left';
+import { DocumentDetailsLeftPanelKey } from '../shared/constants/panel_keys';
 import { useRightPanelContext } from './context';
 
 interface PanelNavigationProps {
