+++ x-pack/plugins/security_solution/public/flyout/document_details/shared/hooks/url/expandable_flyout_state_from_event_meta.ts
@@ -5,7 +5,7 @@
  * 2.0.
  */
 
-import { DocumentDetailsRightPanelKey } from '../../../right';
+import { DocumentDetailsRightPanelKey } from '../../constants/panel_keys';
 
 interface RedirectParams {
   index: string;
